refactor: rm dup code.
